first try of implementing mongodb repo for topics

+++ engine/src/pagination.rs
@@ -1,13 +1,13 @@
 use serde::Deserialize;
 use utoipa::ToSchema;
 
-const fn default_page() -> u32 {
+const fn default_page() -> u64 {
     1
 }
 
 #[derive(Debug, Deserialize, ToSchema, PartialEq, Eq)]
 pub struct Pagination {
     #[serde(default = "default_page")]
-    pub page: u32,
-    pub page_size: Option<u32>,
+    pub page: u64,
+    pub page_size: Option<u64>,
 }
